requirements.txt
environs==9.5.0
Jinja2==3.1.2
python-telegram-bot[rate-limiter]==20.6
langchain==0.0.339
aiogram==3.1.1
openai==1.3.4
beautifulsoup4==4.12.2
selenium==4.15.2
psycopg2==2.9.9
aiohttp
feedparser
python-dateutil